% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyCombMod.R
\name{ApplyCombMod}
\alias{ApplyCombMod}
\title{Apply a combined exponential and linear model}
\usage{
ApplyCombMod(data, y = "y", x = "x")
}
\arguments{
\item{data}{data frame containg x and y data to which the model is ought to be applied to}

\item{y}{name of column in data containing y data}

\item{x}{name of column in data containing x data}
}
\value{
model parameters
}
\description{
a non linear model combining an exponential and a linear fit is applied to the data using the Gauss-Newton algorithm of nls.
starting values are calculated
based on the data. Weights are applied to the model based on the estimated insecurity of the data quality.
}
