% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyCombMod2.R
\name{ApplyCombMod2}
\alias{ApplyCombMod2}
\title{Apply a combined exponential and linear model}
\usage{
ApplyCombMod2(data, y = "y", x = "x")
}
\arguments{
\item{data}{data frame containg x and y data to which combined exponentail and linear model is ought to be applied to}

\item{y}{name of column in data containing y data}

\item{x}{name of column in data containing x data}
}
\value{
model parameters
}
\description{
a non linear model combining an exponential and a linear fit is applied to the data using nls. starting values are calculated
based on the data. Constraints are applied to the model based on the known constraints of the aimed model.
}
