% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFitParam.R
\name{ExtractFitParam}
\alias{ExtractFitParam}
\title{Extracts the fitting parameters from results list}
\usage{
ExtractFitParam(result_list)
}
\arguments{
\item{result_list}{output list from the functions TurgorLossPoint, OsmoticPot, ModElasticity,
StomatalClosure or Gmin}
}
\value{
data frame containing the coefficients and the 0.95 confidence interval of the coefficients from the fit
}
\description{
Extracts the coefficients and confidence intervals from the fitting results of the functions analysing the pressure volume curve
(TurgorLossPoint, OsmoticPot and ModElasticity) or the functions analysing the leaf drying curve
(StomatalClosure and Gmin)
}
