% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeDf.R
\name{MergeDf}
\alias{MergeDf}
\title{Merges data to data frames}
\usage{
MergeDf(x, y, y2 = FALSE, y3 = FALSE, legend, legend.y2 = FALSE,
  legend.y3 = FALSE)
}
\arguments{
\item{x}{vector containing the x values}

\item{y}{vector containing the y values}

\item{y2}{optional vector containing the values for the second y coordinates}

\item{y3}{optional vector containing the values for the third y coordinates}

\item{legend}{name of the y values in the legend}

\item{legend.y2}{optional name of the second y values in the legend}

\item{legend.y3}{optional name of the third y values in the legend}
}
\value{
data frame with columns containing all above information in equalized length as requested by gglot
}
\description{
merges data frames containing all neccessary informations for plotting with PlotOutput()
}
