% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderCheck.R
\name{OrderCheck}
\alias{OrderCheck}
\title{Order Check}
\usage{
OrderCheck(data, sample = FALSE, date.and.time = FALSE,
  fresh.weight = FALSE, water.potential = FALSE,
  time.since.start = FALSE)
}
\arguments{
\item{data}{data frame containing the data to be checked}

\item{sample}{name of the column containing the sample IDs, if present in data}

\item{date.and.time}{name of the column containing the date and time (POSIXct) of the measurements, if present in data}

\item{fresh.weight}{name of the column containing the numeric fresh weight values, if present in data}

\item{water.potential}{name of the column containing the numeric water potential values, if present in data}

\item{time.since.start}{name of the column containing the numeric time since start values, if present in data}
}
\value{
no return value
}
\description{
Checks for the correct ordering of the data: increasing for date.and.time and time.since start, decreasing for fresh.weight and
water.potential. Done separatly for each sample. An individualized warning is printed if not ordered correctly.
}
