% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOutput.R
\name{PlotOutput}
\alias{PlotOutput}
\title{Plot Output}
\usage{
PlotOutput(sub.sample, x, y, y2 = FALSE, y3 = FALSE, legend.y,
  legend.y2 = FALSE, legend.y3 = FALSE, x.axis, y.axis,
  x.intercept = FALSE, y.intercept = FALSE,
  legend.x.intercept = FALSE, line.x, line.y, line.y2 = FALSE,
  line.y3 = FALSE, legend.line.y, legend.line.y2 = FALSE,
  legend.line.y3 = FALSE, show.legend = show.legend)
}
\arguments{
\item{sub.sample}{sample ID}

\item{x}{vector containing the x coordinates of the data to be plotted as points}

\item{y}{vector containing the y coodinates of the data to be plotted as points}

\item{y2}{optional vector containing the second y coordinates of the data to be plotted as points}

\item{y3}{optional vector containing the third y coordinates of the data to be plotted as points}

\item{legend.y}{string, name of data points to be printed in the legend}

\item{legend.y2}{string, optional name of second set of data points to be printed in the legend}

\item{legend.y3}{string, optional name of third set of data points to be printed in the legend}

\item{x.axis}{string, label of x axis}

\item{y.axis}{sring, label of y axis}

\item{x.intercept}{vector containg the x coordinate of the intercept}

\item{y.intercept, }{optional vector containg the y coordinate of the intercept}

\item{legend.x.intercept}{string, name of x.intercept to be printed in the legend}

\item{line.x}{vector containing the x coordinate for the lines}

\item{line.y}{vector containing the y coordinates for the line}

\item{line.y2}{vector containing the y coordinates for the second line}

\item{line.y3}{vector containing the y coordinates for the second line}

\item{legend.line.y}{string, name of line to be printed in the legend}

\item{legend.line.y2}{string, name of second line to be printed in the legend}

\item{legend.line.y3}{string, name of third line to be printed in the legend}

\item{show.legend}{boolean, specifies whether a legend is to be printed}
}
\value{
graphic
}
\description{
plots the data as specified
}
