% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidityCheck.R
\name{ValidityCheck}
\alias{ValidityCheck}
\title{Ensures the validity of the input data}
\usage{
ValidityCheck(data, sample = FALSE, leaf.area = FALSE,
  date.and.time = FALSE, dry.weight = FALSE,
  fresh.weight.saturated = FALSE, fresh.weight = FALSE,
  water.potential = FALSE, RWD = FALSE, conductance = FALSE,
  time.since.start = FALSE, temperature = FALSE, humidity = FALSE)
}
\arguments{
\item{data}{data frame containing the data to be checked}

\item{sample}{name of column containing the sample ID (default: sample)}

\item{leaf.area}{name of column containing the leaf area (cm^2) (default: leaf.area)}

\item{date.and.time}{name of column containing the date and time (default: date and time)}

\item{dry.weight}{name of column containing the dry weight (g) (default: dry weight)}

\item{fresh.weight.saturated}{name of column containing the saturated fresh weight (g) (default: fresh.weight.saturated)}

\item{fresh.weight}{name of column containing the fresh weight (g) (default: fresh.weight)}

\item{water.potential}{name of column containing the water potential (bar) (default: water.potential)}

\item{RWD}{name of column containing the relative water deficit (default: RWD)}

\item{conductance}{name of column containing the conductance values (default: conductance)}

\item{time.since.start}{name of column containing the time since start values (default: time.since.start)}

\item{temperature}{name of column containing the temperature values (default: temperature)}

\item{humidity}{name of column containing the humidity values (default: humidity)}
}
\value{
no return value
}
\description{
Ensures the validity of the input data
}
