% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidityCheckDetail.R
\name{ValidityCheckDetail}
\alias{ValidityCheckDetail}
\title{Checks if column exists in data, is numeric and has the same lenghts as the others existence}
\usage{
ValidityCheckDetail(data_in, value)
}
\arguments{
\item{data_in}{data frame to be checked}

\item{value}{column in data}
}
\value{
no return value
}
\description{
Checks if column exists in data, is numeric and has the same lenghts as the others existence
}
