% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data.R
\docType{data}
\name{weather_data}
\alias{weather_data}
\title{Weather data}
\format{A data frame with 253 rows and 3 variables}
\usage{
weather_data
}
\description{
A dataset containing weather data obtained by a data logger at the site of the measurement of the leaf drying data
(leaf_drying_data)
}
\details{
\itemize{
  \item date.and.time: Time (and date) of the weather logging event (2019-03-25 00:00:00 - 2019-03-27 15:00:00)
  \item temperature: Air temperature in degree Celsius (18.05 - 26.81)
  \item humidity: Relative air humidity in percentage (35.1 - 65.07)
}
}
\keyword{datasets}
